# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Planiranje Putovanja API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/destinations/{destination}": {
            "get": {
                "tags": [
                    "Destinacije"
                ],
                "summary": "Prikaz detalja jedne destinacije",
                "operationId": "d02eefd6431a527f7d131a1174fcfc75",
                "parameters": [
                    {
                        "name": "destination",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalji destinacije"
                    }
                }
            }
        },
        "/api/import/destinations": {
            "post": {
                "tags": [
                    "Import"
                ],
                "summary": "Uvoz destinacija sa eksternog API-ja",
                "operationId": "a52e39a19372242cb5ef7c7bd317e0d4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "query": {
                                        "type": "string",
                                        "example": "Paris"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Uspešan uvoz"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}